backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
pydantic[email]==2.12.4
Jinja2==3.1.4
python-multipart==0.0.9
paramiko==3.4.1
psutil==6.1.0
proxmoxer==2.1.0
requests==2.32.3
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
cryptography>=41.0.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
email-validator==2.2.0
websockets==12.0
apscheduler==3.10.4
